requirements.txt
pyserial
mqtt_codec
schedule